# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API minta backend",
        "version": "1.0.0",
        "description": "<h1>When the impostor is sus ඞ</h1>"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        },
        {
            "url": "https://jedlik.cyclic.app"
        }
    ],
    "security": [{ "session": [] }],
    "paths": {
        "/cities": {
            "get": {
                "summary": "Get all cities",
                "description": "Retrieves all cities from the database",
                "tags": ["City"],
                "responses": {
                    "200": {
                        "description": "Successful response, returns an array of cities",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "An array of cities",
                                    "type": "object",
                                    "properties": {
                                        "cities": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/City"
                                            }
                                        }
                                    }
                                },
                                "example": {
                                    "cities": [
                                        {
                                            "_id": 1,
                                            "code": "BP",
                                            "fullName": "Budapest"
                                        },
                                        {
                                            "_id": 2,
                                            "code": "DE",
                                            "fullName": "Debrecen"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Session id missing or session has expired, please log in!"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": 401
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error occured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": 500
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create new city",
                "description": "Creates a new city",
                "tags": ["City"],
                "parameters": [
                    {
                        "in": "path",
                        "name": "city",
                        "description": "The city object to create",
                        "schema": {
                            "$ref": "#/components/schemas/CreateCityDto"
                        },
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "The city object to create",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCityDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response, returns the created city",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                },
                                "example": {
                                    "_id": 1,
                                    "code": "BP",
                                    "fullName": "Budapest"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Session id missing or session has expired, please log in!"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": 401
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error occured",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "number",
                                            "example": 500
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateCityDto": {
                "type": "object",
                "description": "The class that contains the data for creating a new city.",
                "properties": {
                    "_id": {
                        "type": "number",
                        "format": "int32"
                    },
                    "code": {
                        "type": "string",
                        "description": "The code of the city, max 2 characters, Hungarian characters.",
                        "example": "BP",
                        "maxLength": 2
                    },
                    "fullName": {
                        "type": "string",
                        "description": "The full name of the city, max 50 characters, Hungarian characters.",
                        "example": "Budapest",
                        "maxLength": 50
                    }
                },
                "required": ["code", "fullName"]
            },
            "City": {
                "type": "object",
                "description": "Database object that stores information about a city.",
                "properties": {
                    "_id": {
                        "type": "number",
                        "description": "The city's id"
                    },
                    "code": {
                        "type": "string",
                        "description": "The city's code",
                        "maxLength": 2
                    },
                    "fullName": {
                        "type": "string",
                        "description": "The full name of the city",
                        "maxLength": 50
                    }
                },
                "required": ["_id", "code", "fullName"],
                "additionalProperties": false
            },
...
        },
        "securitySchemes": {
            "session": {
                "type": "apiKey",
                "in": "cookie",
                "name": "mj-session"
            }
        }
    }
}